sdlc(impl): Extract nextPause + backoff constants to scripts/backoff.ts
- Files changed: scripts/backoff.ts (new — exports nextPause, MIN_PAUSE_SEC, MAX_PAUSE_SEC, BACKOFF_FACTOR), scripts/self_runner.ts (remove local defs, add import), scripts/loop_in_claude.ts (remove local defs, add import)
- Tests: scripts/backoff_test.ts (new — moved 3 nextPause tests), scripts/self_runner_test.ts (removed nextPause import + 3 tests)
- Check: pending
